create jabatan tunjangan potongan

+++ resources/views/admin/permissions/index.blade.php
@@ -12,9 +12,11 @@
   <div class="row">
     <div class="col-xs-12">
       <div class="box">
-        <!-- <div class="box-header">
-          <h3 class="box-title">Data Table With Full Features</h3>
-        </div> -->
+      @can('permission-create')
+        <div class="box-header my-3">
+            <a href="permissions/create" class="btn btn-primary btn-sm">Create Permission</a>
+        </div>
+      @endcan 
         <div class="box-body">
           <table id="example1" class="table table-bordered table-hover">
             <thead>
@@ -22,7 +24,9 @@
                 <th>Id</th>
                 <th>Name</th>
                 <th>Guard Name</th>
-                <th>Action</th>
+                @if(auth()->user()->hasAnyPermission(['permission-edit','permission-delete']))
+                    <th>Action</th> 
+                @endif
               </tr>
             </thead>
             <tbody>
@@ -31,19 +35,24 @@
                   <td><?= $data->id; ?> </td>
                   <td><?= $data->name; ?> </td>
                   <td><?= $data->guard_name; ?> </td>
+                  @if(auth()->user()->hasAnyPermission(['permission-edit','permission-delete']))
                   <td>
                     <div class="row">
+                    @can('permission-edit')
                       <div class="col">
                         <a class="btn btn-primary btn-sm" href="{{route('permissions.edit',$data->id)}}">edit</a>
                       </div>
+                    @endcan
+                    @can('permission-delete')
                       <div class="col">
-                        <!-- <button class="btn btn-warning btn-sm">delete</button> -->
                            {!! Form::open(['method' => 'DELETE','route' => ['permissions.destroy', $data->id],'style'=>'display:inline']) !!}
                                 {!! Form::submit('Delete', ['class' => 'btn btn-warning btn-sm']) !!}
                             {!! Form::close() !!}
                       </div>
+                    @endcan
                     </div>
                   </td>
+                  @endif
                 </tr>
               @endforeach
             </tbody>
